Set grid dimensions to integers
The amout of cells in the grid will be width*height.

+++ src/TextGrid.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 #include <string>
 
-TextGrid::TextGrid(double height, double width) {
+TextGrid::TextGrid(int height, int width) {
   m_height = height;
   m_width  = width;
 }
